+++ NEWS.md
@@ -6,6 +6,7 @@
 * Updated examples and tests using `scda` synthetic data to replace `random.cdisc.data`.
 * Added `filter_spec` support in  `tm_outliers` module.
 * Added ordered select input support in  `tm_g_scatterplotmatrix`, `tm_g_association` and `tm_t_crosstable` modules.
+* Added validations to the theoretical distributions' parameters of `tm_g_distribution`.
 
 ### Miscellaneous
 * Removed mentions of Roche internal services from the README file.